Check that all structural formula PNGs have been created

Checks that PNGs for structural formulae are consistent with the compounds in the spectral library. It compares the InChI-Keys in the library and filenames and prints any inconsistencies to the console.

Usage

checkPngAvailability(databasePath, targetDir)

Arguments

databasePath

Path to spectral library (CSL, SQLite format)

targetDir

Path where the PNGs are to be saved