Ingest featureRecords into Elasticsearch

The screening processing output is an RDS file for each batch (or more than one for large batches). These are ingested into NTSPortal Elasticsearch.

Usage

ingest(path, ingestPipeline = "ingest-feature")

Arguments

path

Path to single RDS file or directory with RDS files.

ingestPipeline

ID of ingest pipeline to use (default is "ingest-feature" pipeline)

Value

list with one top-level entry per batch (RDS file). The second level has one entry per alias and this contains a character vector of the index name associated with that alias.

Details

Ingest pipelines Feature documents are enriched with sample and analytical method information using an ingest pipeline. Use `ingestPipeline` to specify which ingest pipeline to use (must have been previously submitted to Elasticsearch)